xfst/lexc rewrite
This is a complete xfst/lexc rewrite of SFST sources. The
intention is to drop the SFST version in time, and only
maintain this branch. For now, there is nothing exciting,
the analyzer should be almost the same as the current SFST
version on the master branch. However, there are some
differences (which may be bad), and this version is not
tested and documented as good as the SFST version. On the
other hand, some wrinkles are ironed out, and there are
already some gems in place which may be useful for future
improvements.

This version compiles fine with foma. It should probably
work fine with HFST too, but it is not yet tested.